Hide KeyClear feature
Changes added : commented code where KeyClear option is added

Problem : KeyClear option is IBM feature and not required for others

Solution : Commented code where KeyClear option is added.

Since the plan was to hide it but, not to remove it.

Testing : Tested webui with the changes and verified that KeyClear option is not visible.

Fixes nvbug https://nvbugs/4042790