Remove phosphor-rest style login
As part of [1] There are deprecated login flows that webui-vue seems to
have copied from phosphor-rest.  These were originally added because
phosphor-rest didn't look at response codes, but webui-vue does, so we
can use the normal version, which allows us to reduce the code in
bmcweb.

This needs to go the same time as [1].

Tested: Gunnar tested with 65810 and this works.

[1] https://gerrit.openbmc.org/c/openbmc/bmcweb/+/65810

Change-Id: I1e55f08cf1d7d44b6757ac4173a26546eaca72e5
Signed-off-by: Ed Tanous <edtanous@google.com>
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>