Fix Redfish logo on automatic logout after inactivity
```
Changes added : Modified to code to access the "/" window instead of "/login"

Problem : After 30 minutes in-activity the session get logged out and when user tries to access the web it shows redfish logo as below snapshot.

Solution : Modified the code to access the "/" window instead of "/login" window. So that when user tries to access the web it will redirect to login page instead of redfish logo page.

Testing:
1.	Login to the web and wait for 30 minutes.
2.	After 30 minutes try to access the web and it will redirect to login page instead of redfish logo page.

Snapshot :

```
Fixes nvbug https://nvbug/4194097