hide backup card if backup image is empty
[Problem]: Backup card gets displayed on UI even if backup image
is not available on BMC or Host.

[Solution]: Show backup card view only when backup firmware image is
available.

[Logic]: All bmc and host firmwares are stored in bmcSoftwareImageIds and
hostSoftwareImageIds. If there are more than one image, backupBmcFirmware
and backupHostFirmware will not be empty since only one active image
at a time. So, using back() function which fetches backupBmcFirmware
and backupHostFirmware array from store as a checker to display
BMC/Host Backup card.
As part of this change, hide "Switch to running" button if backup image is
available but not "updateable".

Fixes nvbug https://nvbugswb.nvidia.com/NVBugs5/redir.aspx?url=/4630378

Signed-off-by: Shubhi Garg <shgarg@nvidia.com>