inventory: fix quicklinks to show valid inventory
[Problem]:
Some of the inventory components (Ex DIMM, Fans) might not render
if response from backend is NULL. But "Quicklinks" still shows all
the components. On clicking to such components from "Quicklinks" takes
us to nowhere. WebUI debug console throws error:
TypeError: e.setAttribute is not a function

[Goal]:
There should not be references in "Quicklinks" for components which are
not visible in Inventory page.

[Solution]:
Once UI DOM elements are mounted, check all the links from $refs if they are
pointing to a valid reference and store in validLinks array. If a link is
not valid, it will point to a "COMMENT" type node. Use "validLinks" array
in "Quicklinks" template to render links.

Fixes Nvbug https://nvbugspro.nvidia.com/bug/4747379

Signed-off-by: Shubhi Garg <shgarg@nvidia.com>