igx: hide security and access section
Problem: Security and access section is not working properly
and right now it is not getting used for igx platform
Users try to access security section while it is not working.

Solution: Hide the complete security and access section and
add it later when it gets completely fixed from backend.

Changes: Removed all the modules of security and access from
nvidia-igx.js file in env folder.

Fixes nvbug https://nvbugspro.nvidia.com/4576015

Signed-off-by: Ishan Verma ishanv@nvidia.com