fix start update firmware message
[Problem]:
When click on "Start update" button, UI pops up a message:
"The BMC may reboot during the update process. The host cannot be
powered on until the update is finished.
Are you sure you want to proceed with the update?"

This message makes the assumption as if BMC firmware is getting updated
which is not right. We can update any target modules from UI.
Currently, "isSingleFileUploadEnabled" is used as a condition to render
bmc fw update message. This condition will be true if there is no BIOS
in the system.

If there is a BIOS with backup image, then current message does not
support showing backup host(bios) version.

[Solution]:
Fix message to make it generic for modules other than BMC/BIOS.
Do not rely on "isSingleFileUploadEnabled" condition check.

[Changes]:
Added checks for BMC backup and BIOS backup. If any of these backups are
present then show backup message else keep a default message.

Fixes Nvbug https://nvbugspro.nvidia.com/4592117

Signed-off-by: Shubhi Garg <shgarg@nvidia.com>